+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased] - XX-XX-XX
 
+### Changed
+
+- Improve code quality according to codacy checkup
+
 # [3.6.2] - 2019-04-24
 
 ### Added